packages/core/scripts/web-search/requirements.txt
requests>=2.31.0
tavily-python>=0.3.0
